backend/requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0

# File Processing
python-multipart>=0.0.6
pymupdf>=1.23.0
python-docx>=1.1.0

# AI/ML
httpx>=0.26.0
chromadb>=0.4.22
tiktoken>=0.5.2

# Security
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0
slowapi>=0.1.9

# File Validation
python-magic-bin>=0.4.14  # Windows; use python-magic==0.4.27 in Docker

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.0
